# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API cho Trang thông tin điện tử Viện Công nghệ số và Chuyển đổi số quốc gia (NIDIT) - bản trình diễn
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /admin/stats/visits:
    get:
      operationId: getVisitStats
      tags: [admin]
      parameters:
        - name: days
          in: query
          required: false
          schema: { type: integer, minimum: 1, maximum: 365 }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/VisitStats" }
  /admin/activity-logs:
    get:
      operationId: listActivityLogs
      tags: [admin]
      parameters:
        - $ref: "#/components/parameters/PageQuery"
        - $ref: "#/components/parameters/PageSizeQuery"
        - name: actorId
          in: query
          required: false
          schema: { type: integer }
        - name: entityType
          in: query
          required: false
          schema: { type: string }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ActivityLogPage" }

  # ---------------------------------------------------------------- users
  /admin/users:
    get:
      operationId: listUsers
      tags: [admin]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/User" }
    post:
      operationId: createUser
      tags: [admin]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/UserInput" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/User" }
  /admin/users/{id}:
    parameters:
      - $ref: "#/components/parameters/IdPath"
    patch:
      operationId: updateUser
      tags: [admin]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/UserUpdate" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/User" }

  # ---------------------------------------------------------------- backups
  /admin/backups:
    get:
      operationId: listBackups
      tags: [admin]
      summary: Danh sách bản sao lưu (tải về tại /api/admin/backups/{id}/download?token=...)
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Backup" }
    post:
      operationId: createBackup
      tags: [admin]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/BackupInput" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Backup" }

  # ---------------------------------------------------------------- media library
  /admin/media:
    get:
      operationId: listMedia
      tags: [admin]
      parameters:
        - name: type
          in: query
          required: false
          schema: { $ref: "#/components/schemas/MediaType" }
        - name: q
          in: query
          required: false
          schema: { type: string }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/MediaItem" }
    post:
      operationId: createMedia
      tags: [admin]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/MediaInput" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/MediaItem" }
  /admin/media/{id}:
    parameters:
      - $ref: "#/components/parameters/IdPath"
    delete:
      operationId: deleteMedia
      tags: [admin]
      responses:
        "204":
          description: Deleted
components:
  parameters:
    IdPath:
      name: id
      in: path
      required: true
      schema: { type: integer }
...
    PageQuery:
      name: page
      in: query
      required: false
      schema: { type: integer, minimum: 1 }
    PageSizeQuery:
      name: pageSize
      in: query
      required: false
      schema: { type: integer, minimum: 1, maximum: 100 }
...
  schemas:
...
    UserRole:
      type: string
      enum: [admin, editor, reviewer]
...
    MediaType:
      type: string
      enum: [image, video, file]

    # ------------------------------------------------------------ shared small types
...
    LabelCount:
      type: object
      required: [label, count]
      properties:
        label: { type: string }
        count: { type: integer }
...
    User:
      type: object
      required: [id, fullName, email, role, title, unit, isActive, lastLoginAt, createdAt]
      properties:
        id: { type: integer }
        fullName: { type: string }
        email: { type: string }
        role: { $ref: "#/components/schemas/UserRole" }
        title: { type: string, description: Chức danh }
        unit: { type: string, description: Đơn vị công tác }
        isActive: { type: boolean }
        lastLoginAt: { type: ["string", "null"], format: date-time }
        createdAt: { type: string, format: date-time }
    UserInput:
      type: object
      required: [fullName, email, role]
      properties:
        fullName: { type: string, minLength: 1 }
        email: { type: string, minLength: 3 }
        role: { $ref: "#/components/schemas/UserRole" }
        title: { type: string }
        unit: { type: string }
    UserUpdate:
      type: object
      properties:
        fullName: { type: string, minLength: 1 }
        email: { type: string, minLength: 3 }
        role: { $ref: "#/components/schemas/UserRole" }
        title: { type: string }
        unit: { type: string }
        isActive: { type: boolean }
...
    DailyVisit:
      type: object
      required: [date, views, visitors]
      properties:
        date: { type: string, description: "YYYY-MM-DD" }
        views: { type: integer }
        visitors: { type: integer }
    PathCount:
      type: object
      required: [path, views]
      properties:
        path: { type: string }
        views: { type: integer }
    ArticleViewCount:
      type: object
      required: [id, title, slug, views]
      properties:
        id: { type: integer }
        title: { type: string }
        slug: { type: string }
        views: { type: integer }
    VisitStats:
      type: object
      required: [days, totalViews, totalVisitors, avgViewsPerDay, daily, topPages, topArticles, devices, referrers, sections]
      properties:
        days: { type: integer }
        totalViews: { type: integer }
        totalVisitors: { type: integer }
        avgViewsPerDay: { type: integer }
        daily:
          type: array
          items: { $ref: "#/components/schemas/DailyVisit" }
        topPages:
          type: array
          items: { $ref: "#/components/schemas/PathCount" }
        topArticles:
          type: array
          items: { $ref: "#/components/schemas/ArticleViewCount" }
        devices:
          type: array
          items: { $ref: "#/components/schemas/LabelCount" }
        referrers:
          type: array
          items: { $ref: "#/components/schemas/LabelCount" }
        sections:
          type: array
          description: Lượt xem theo chuyên trang
          items: { $ref: "#/components/schemas/LabelCount" }
    ActivityLog:
      type: object
      required: [id, actorId, actorName, actorRole, action, actionLabel, entityType, entityId, entityTitle, detail, createdAt]
      properties:
        id: { type: integer }
        actorId: { type: ["integer", "null"] }
        actorName: { type: string, description: "Tên người thực hiện hoặc 'Hệ thống'" }
        actorRole: { type: ["string", "null"] }
        action: { type: string }
        actionLabel: { type: string, description: Mô tả thao tác bằng tiếng Việt }
        entityType: { type: string }
        entityId: { type: ["integer", "null"] }
        entityTitle: { type: ["string", "null"] }
        detail: { type: ["string", "null"] }
        createdAt: { type: string, format: date-time }
    ActivityLogPage:
      type: object
      required: [items, total, page, pageSize, totalPages]
      properties:
        items:
          type: array
          items: { $ref: "#/components/schemas/ActivityLog" }
        total: { type: integer }
        page: { type: integer }
        pageSize: { type: integer }
        totalPages: { type: integer }
...
    BackupTableCount:
      type: object
      required: [table, rows]
      properties:
        table: { type: string }
        rows: { type: integer }
    Backup:
      type: object
      required: [id, note, createdByName, createdAt, sizeBytes, tables]
      properties:
        id: { type: integer }
        note: { type: ["string", "null"] }
        createdByName: { type: ["string", "null"] }
        createdAt: { type: string, format: date-time }
        sizeBytes: { type: integer }
        tables:
          type: array
          items: { $ref: "#/components/schemas/BackupTableCount" }
    BackupInput:
      type: object
      properties:
        note: { type: ["string", "null"] }

    # ------------------------------------------------------------ media
    MediaItem:
      type: object
      required: [id, url, title, alt, type, createdAt]
      properties:
        id: { type: integer }
        url: { type: string }
        title: { type: string }
        alt: { type: ["string", "null"] }
        type: { $ref: "#/components/schemas/MediaType" }
        createdAt: { type: string, format: date-time }
    MediaInput:
      type: object
      required: [url, title]
      properties:
        url: { type: string, minLength: 1 }
        title: { type: string, minLength: 1 }
        alt: { type: ["string", "null"] }
        type: { $ref: "#/components/schemas/MediaType" }